requirements.txt
streamlit>=1.30.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.15
easyocr>=1.7.1
PyMuPDF>=1.23.8
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pdfplumber>=0.10.3
numpy>=1.24.0
pandas>=2.1.0
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"
